% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollr_analyze.R
\name{pollr_analyse}
\alias{pollr_analyse}
\title{Analyze a survey question and return summary data, table and interactive plot}
\usage{
pollr_analyse(survey_data, question_varname, weight_varname = NULL)
}
\arguments{
\item{survey_data}{A dataframe with survey data}

\item{question_varname}{A character string, the variable name to analyze}

\item{weight_varname}{A character string, the variable name for weights (optional)}
}
\value{
A list with three elements:
\describe{
\item{data}{A tibble of data question}
\item{results}{A tibble of data question results}
\item{table}{An HTML table of the data question results}
\item{plot}{An interactive plot of the data question results}
}
}
\description{
Analyze a survey question and return summary data, table and interactive plot
}
